cleared trainingpage from react media and styles
