+++ src/components/StatisticsData/StatisticsData.module.scss
@@ -1,0 +1,31 @@
+@import "../../assets/sass/main.scss";
+.page__wrapperStat {
+  display: flex;
+  flex-direction: column;
+  padding-top: 15px;
+  padding-bottom: 30px;
+  @include vp(tablet) {
+    padding-top: 30px;
+  }
+  @include vp(desktop) {
+    flex-direction: row-reverse;
+    padding-top: 15px;
+  }
+}
+.left__wrapper {
+  margin-bottom: 30px;
+  @include vp(desktop) {
+    border: 1px solid green;
+    margin-right: 40px;
+    margin-bottom: 0;
+  }
+}
+.right__wrapper {
+  margin-bottom: 30px;
+  @include vp(tablet) {
+    margin-bottom: 0;
+  }
+  @include vp(desktop) {
+    border: 1px solid red;
+  }
+}
