--- src/pages/TrainingPage/TrainingPage.module.scss
@@ -1,76 +1,0 @@
-@import "../../assets/sass/main.scss";
-.page__wrapper {
-  display: flex;
-  flex-direction: column;
-  padding-top: 30px;
-  padding-bottom: 30px;
-  @include vp(desktop) {
-    flex-direction: row-reverse;
-    padding-top: 50px;
-  }
-}
-.page__wrapperStat {
-  display: flex;
-  flex-direction: column;
-  padding-top: 15px;
-  padding-bottom: 30px;
-  @include vp(tablet) {
-    padding-top: 30px;
-  }
-  @include vp(desktop) {
-    flex-direction: row-reverse;
-    padding-top: 15px;
-  }
-}
-.left__wrapper {
-  margin-bottom: 30px;
-  @include vp(desktop) {
-    border: 1px solid green;
-    margin-right: 40px;
-    margin-bottom: 0;
-  }
-}
-.right__wrapper {
-  margin-bottom: 30px;
-  @include vp(tablet) {
-    margin-bottom: 0;
-  }
-  @include vp(desktop) {
-    border: 1px solid red;
-  }
-}
-.button__plus {
-  position: fixed;
-  // right: 136px;
-  // bottom: 25px;
-  bottom: 0;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  width: 52px;
-  height: 52px;
-  border: none;
-  border-radius: 50%;
-  background-color: var(--accent-active-cl);
-  &:hover,
-  &:focus {
-    background-color: var(--accent-active-cl-hover);
-  }
-}
-
-.react__icon {
-  color: var(--main-bg-cl);
-  position: relative;
-  top: 0;
-  right: 0;
-}
-.button__arrow {
-  border: none;
-  background-color: transparent;
-  width: 30px;
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-}
-.icon__arrow {
-  color: var(--accent-active-cl);
-}
